requirements.txt
groq==0.13.1
requests==2.31.0
beautifulsoup4==4.12.3
feedparser==6.0.11
apscheduler==3.10.4
python-dotenv==1.0.1
